requirements_reddit.txt
# Reddit Scraping System Dependencies

# Core Reddit API
praw>=7.7.0

# Database
pymongo>=4.3.3

# Task Queue
celery>=5.3.0
redis>=4.5.0

# Scheduling
schedule>=1.2.0

# Environment Management
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Logging and Monitoring
flower>=2.0.0  # Celery monitoring

# Optional: For advanced analytics
# scikit-learn>=1.3.0
# matplotlib>=3.7.0
# seaborn>=0.12.0 